source/dataagent-harbor/pyproject.toml
[project]
name = "dataagent-harbor"
version = "0.1.0"
description = "Testing and evaluation framework for DataAgent Server"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "DataAgent Team"}
]
keywords = ["agent", "testing", "benchmark", "evaluation", "tracing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "dataagent-core",
    "dataagent-server",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "langsmith>=0.4.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "hypothesis>=6.0.0",
]

[project.scripts]
dataagent-harbor = "dataagent_harbor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dataagent_harbor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
dataagent-core = { path = "../dataagent-core" }
dataagent-server = { path = "../dataagent-server" }
